Add files via upload
Retail analysis of superstore data wherein we find out what are the sales ratio in different segments and based on product variant also we determine the profit difference based on region and also do a demography analysis to understand the performance of superstore sales across the region. Which helps us to define the improvement factor for business.